BUG 3323 - Changing sql queries to prepared statements in idmlite
By-hand cherry-pick of https://git.opendaylight.org/gerrit/#/c/22051/2.

+++ aaa-idmlight/src/main/java/org/opendaylight/aaa/idm/IdmLightProxy.java
@@ -160,9 +160,9 @@ public class IdmLightProxy implements CredentialAuth<PasswordCredentials>,
     public List<String> listDomains(String userId) {
         debug("list Domains for userId:" + userId);
         List<String> domains = new ArrayList<String>();
-        long uid=0;
+        int uid=0;
         try {
-           uid = Long.parseLong(userId);
+           uid = Integer.parseInt(userId);
         }
         catch (NumberFormatException nfe) {
            logger.warn("not a valid userid:" + userId);
@@ -198,12 +198,12 @@ public class IdmLightProxy implements CredentialAuth<PasswordCredentials>,
               debug("DomainName: " + domainName + " Not found!");
               return roles;
            }
-           long did = domainList.get(0).getDomainid();
+           int did = domainList.get(0).getDomainid();
 
            // validate userId
-           long uid=0;
+           int uid=0;
            try {
-              uid = Long.parseLong(userId);
+              uid = Integer.parseInt(userId);
            }
            catch (NumberFormatException nfe) {
               logger.warn("not a valid userid:" + userId);
